Apply workaround only on 26.04+

+++ entrypoint.sh
@@ -113,10 +113,13 @@ fi
 if grep -q securityfs /proc/filesystems; then
     mount -o rw,nosuid,nodev,noexec,relatime securityfs -t securityfs /sys/kernel/security
 fi
-if [ -f /sys/module/apparmor/parameters/enabled ]; then
+
+source /etc/os-release
+if [ "${VERSION_ID%%.*}" -ge 26 ] && [ -f /sys/module/apparmor/parameters/enabled ]; then
   echo N > /tmp/apparmor_disabled
   mount --bind /tmp/apparmor_disabled /sys/module/apparmor/parameters/enabled 2>/dev/null || true
 fi
+
 mkdir -p /run/lock
 mount -t tmpfs tmpfs /run
 mount -t tmpfs tmpfs /run/lock
